Setup basic project structure
Added typescript as a dependency for this project since
we are entirely based on external api for data.

Added some basic settings and dev-setup guides. We should
update dev-setup guide as we go in the future.

Added a basic hello world template

Closes #1

Revert "Setup basic project structure"

This reverts commit 91013917fb40c47211d01a29b34f12826693e340.

Setup basic project structure

Added typescript as a dependency for this project since
we are entirely based on external api for data.

Added some basic settings and dev-setup guides. We should
update dev-setup guide as we go in the future.

Added a basic hello world template

Closes #1