4 - Added main page
Added main page details to populate vin

Added textfield to display title and other display information. Also
added form fields to support VIN number and Year of Manufacture. Added
submit button to populate data even though for now, we only populate a test
data script.

Added axios as a http request handler to support api fetch calls.

Params were added in the api call as well. We can future proof it with
proper param handling but this should work for now since we have test data
anyways. Created ticket #6 to handle this issue.

Added ability to view data for specific Vehicle

While working on this, we had a lot of back and forth on trying to find the
best layout structure for a lot of data. Granted that table would be the best
of all, but we wanted to see what others had in mind. For now, we are just going
to keep it simple like a card layout with two columns.

Closes #4